backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.1
pydantic-settings==2.8.1
PyJWT==2.10.1
openai==1.68.2
psycopg[binary]==3.2.6
pytest==8.3.5
httpx==0.28.1